Move megaraid plugin
Move the local plugin files to $(pythondir)/local_plugin
ref. https://github.com/libstorage/libstoragemgmt/issues/440
