+++ plugin/Makefile.am
@@ -1,4 +1,4 @@
-SUBDIRS=simc megaraid hpsa_plugin arcconf_plugin nfs_plugin local_plugin smispy targetd_plugin sim_plugin
+SUBDIRS=simc megaraid_plugin hpsa_plugin arcconf_plugin nfs_plugin local_plugin smispy targetd_plugin sim_plugin
 
 plugindir = $(pythondir)/lsm/plugin
 plugin_PYTHON= __init__.py
