--- plugin/megaraid/__init__.py
@@ -1,1 +1,0 @@
-from lsm.plugin.megaraid.megaraid import MegaRAID
