+++ plugin/megaraid_plugin/__init__.py
@@ -1,0 +1,1 @@
+from megaraid_plugin.megaraid import MegaRAID
