+++ plugin/megaraid_plugin/megaraid_lsmplugin.in
@@ -22,14 +22,14 @@ import traceback
 
 try:
     from lsm import PluginRunner
-    from lsm.plugin.megaraid import MegaRAID
+    from megaraid_plugin import MegaRAID
 
     if __name__ == '__main__':
         PluginRunner(MegaRAID, sys.argv).run()
 except Exception:
-    #This should be quite rare, but when it does happen this is pretty
-    #key in understanding what happened, especially when it happens when
-    #running from the daemon.
+    # This should be quite rare, but when it does happen this is pretty
+    # key in understanding what happened, especially when it happens when
+    # running from the daemon.
     msg = str(traceback.format_exc())
     syslog.syslog(syslog.LOG_ERR, msg)
     sys.stderr.write(msg)
